Restructured the boilerplate
In the previous structure, setting up a fresh boilerplate was a little cumbersome, as it involved having to delete the boilerplate's own `.git` folder. With this new structure, rather than deleting the boilerplate's `.git` folder, we just need to make a fresh copy of the `dev` folder, then we are free to immediately initialise a new instance of Git in the folder.